+++ pkg/otel/metrics/arrow/exemplar.go
@@ -21,7 +21,7 @@ import (
 	"math"
 	"sort"
 
-	"github.com/apache/arrow/go/v12/arrow"
+	"github.com/apache/arrow/go/v14/arrow"
 	"go.opentelemetry.io/collector/pdata/pmetric"
 
 	carrow "github.com/open-telemetry/otel-arrow/pkg/otel/common/arrow"